# docs/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: api
  description: ''
  license:
    name: ''
  version: 0.1.0
servers:
- url: /api
paths:
...
  /fitness:
    get:
      tags:
      - fitness
      operationId: get_fitness_freshness
      parameters:
      - name: start_date
        in: query
        required: false
        schema:
          type:
          - string
          - 'null'
        example: 2026-09-20
      - name: end_date
        in: query
        required: false
        schema:
          type:
          - string
          - 'null'
        example: 2026-09-27
      responses:
        '200':
          description: Daily fitness, fatigue, and form data for the authenticated user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FitnessFreshnessResponse'
              example:
                end_date: 2026-09-27
                fatigue_window_days: 7
                fitness_window_days: 42
                points:
                - date: 2026-09-26
                  fatigue: 56.1
                  fitness: 48.2
                  form: -7.9
                  training_load: 72.5
                start_date: 2026-09-20
        '400':
          description: Invalid query parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '401':
          description: Not authenticated
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
      security:
      - bearer_auth: []
...
components:
  schemas:
...
    ApiErrorResponse:
      type: object
      required:
      - message
      properties:
        errors:
          type:
          - object
          - 'null'
          additionalProperties:
            type: array
            items:
              type: string
          propertyNames:
            type: string
          example:
            title:
            - Title cannot be empty
        message:
          type: string
          example: Invalid request
        retry_at:
          type:
          - string
          - 'null'
          format: date-time
      example:
        errors:
          title:
          - Title cannot be empty
        message: Invalid request
...
    FitnessFreshnessPoint:
      type: object
      required:
      - date
      - training_load
      - fitness
      - fatigue
      - form
      properties:
        date:
          type: string
          example: 2026-09-26
        fatigue:
          type: number
          format: double
          example: 56.1
        fitness:
          type: number
          format: double
          example: 48.2
        form:
          type: number
          format: double
          example: -7.9
        training_load:
          type: number
          format: double
          example: 72.5
      example:
        date: 2026-09-26
        fatigue: 56.1
        fitness: 48.2
        form: -7.9
        training_load: 72.5
    FitnessFreshnessResponse:
      type: object
      required:
      - start_date
      - end_date
      - fitness_window_days
      - fatigue_window_days
      - points
      properties:
        end_date:
          type: string
          example: 2026-09-27
        fatigue_window_days:
          type: integer
          format: int32
          example: 7
        fitness_window_days:
          type: integer
          format: int32
          example: 42
        points:
          type: array
          items:
            $ref: '#/components/schemas/FitnessFreshnessPoint'
          example:
          - date: 2026-09-26
            fatigue: 56.1
            fitness: 48.2
            form: -7.9
            training_load: 72.5
        start_date:
          type: string
          example: 2026-09-20
      example:
        end_date: 2026-09-27
        fatigue_window_days: 7
        fitness_window_days: 42
        points:
        - date: 2026-09-26
          fatigue: 56.1
          fitness: 48.2
          form: -7.9
          training_load: 72.5
        start_date: 2026-09-20
...
  securitySchemes:
    bearer_auth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...